requirements.txt
# ABC AI - Agent Bear Corps Runtime
# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
pyyaml>=6.0
cryptography>=41.0.0

# LLM Provider SDKs
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0

# Optional: For enhanced capabilities
requests>=2.31.0
rich>=13.0.0